[refactor/#13]  좌석 등록뷰 파일 분리
